Refactor file and class logic
